fix(lint): clear golangci findings in linux-platform files
Linting on a linux host surfaces golangci findings in //go:build linux
files that a non-linux dev host build-excludes and never lints, so the
linux-platform code had accumulated 81 findings. This fixes 62 in code:
errcheck (proper error handling / slog logging in syscall cleanup),
gocritic (filepathJoin, paramTypeCombine, assignOp, hugeParam,
appendAssign), goconst, noctx (context-aware net APIs), unconvert,
unparam, unused (dead code removed), nilnil, staticcheck. Also converts
4 British spellings to US to match the codebase convention (neighbor
1833:3, behavior 423:0). The residual 18 findings (15 gosec
syscall-inherent, 3 nilerr plugin-Response pattern) are deferred to
plan/spec-lint-linux-residual.md for a security-review / nolint-policy
decision. Verified: go build ./... clean; scoped golangci-lint shows
only the deferred residual.

+++ internal/plugins/ntp/clock_linux.go
@@ -28,7 +28,7 @@ func slewClock(offset time.Duration) error {
 	usec := offset.Microseconds()
 	tx := &syscall.Timex{
 		Modes:  0x0001, // ADJ_OFFSET
-		Offset: int64(usec),
+		Offset: usec,
 	}
 	if _, err := syscall.Adjtimex(tx); err != nil {
 		return fmt.Errorf("adjtimex: %w", err)
